% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{bb}
\alias{bb}
\title{An abstract representation of a brainbrowser widget}
\usage{
bb(obj = NULL, intensities = NULL, color_map = NULL, min = NULL,
  max = NULL, zoom = NULL, rotation = NULL, matrix = NULL,
  debug = FALSE, bg_colour = "", bg_plot = NULL, width = 1,
  height = 1)
}
\arguments{
\item{obj}{The object to display}

\item{intensities}{Intensity values, one per mesh vertex}

\item{color_map}{color values, one per mesh vertex}

\item{min}{minimum for the displayed intensities}

\item{max}{maximum for the displayed intensities}

\item{debug}{whether to append the viewer object to
the html window (no longer necessary, the viewer is attached
to the widget dom element).}

\item{bg_colour}{The background colour for the widget, an empty
string has a clear background.}

\item{bg_plot}{A command to produce a plot in the background of a widget}
}
\description{
This is used to construct layouts containing multiple widgets.
Primarily just stores the arguments to evaluate at render time.
}
