% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{bbcol}
\alias{bbcol}
\title{Abstract column layout for brainbrowser widgets.}
\usage{
bbcol(old, ..., height = 1, width = 1, bg_plot = NULL, data = NULL)
}
\arguments{
\item{old}{A layout object `bb`, `bbrow`, or `bbcol`}

\item{...}{More unnamed layout objects}

\item{height}{A height for the column}

\item{width}{A width for the column}

\item{data}{A list containing arguments to `bb` to be applied where child layout
objects are missing arguments. This can be used to set the same mesh for the whole
column for example.}
}
\description{
Takes one or more layout objects (bb, bbrow, and bbcols) and assembles them
into a column. Data can be passed in via the `data` argument, this can be used to
fill missing data into the child layout objects.
}
