% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{bbrow}
\alias{bbrow}
\title{Abstract row layout for brainbrowser widgets.}
\usage{
bbrow(old, ..., height = 1, width = 1, bg_plot = NULL, data = NULL)
}
\arguments{
\item{old}{A layout object `bb`, `bbrow`, or `bbcol`}

\item{...}{More unnamed layout objects}

\item{height}{A height for the row}

\item{width}{A width for the row}

\item{data}{A list containing arguments to `bb` to be applied where child layout
objects are missing arguments. This can be used to set the same mesh for the whole
row for example.}
}
\description{
Takes one or more layout objects (bb, bbrow, and bbcols) and assembles them
into a row. Data can be passed in via the `data` argument, this can be used to
fill missing data into the child layout objects. The width is
divided amongst the children by their relative widths.
}
