% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{brainbrowser}
\alias{brainbrowser}
\title{Generate a brainbrowser html widget of a given obj}
\usage{
brainbrowser(obj, intensities = NULL, color_map = NULL, min = NULL,
  max = NULL, zoom = 1, rotation = NULL, matrix = NULL,
  debug = FALSE, bg_colour = "", bg_plot = NULL, width = "100\%",
  height = "400px")
}
\arguments{
\item{obj}{The object to display}

\item{intensities}{Intensity values, one per mesh vertex}

\item{color_map}{color values, one per mesh vertex}

\item{min}{minimum for the displayed intensities}

\item{max}{maximum for the displayed intensities}

\item{debug}{whether to append the viewer object to
the html window (no longer necessary, the viewer is attached
to the widget dom element).}

\item{bg_colour}{The background colour for the widget, an empty
string has a clear background.}

\item{bg_plot}{A command to produce a plot in the background of a widget}

\item{A}{zoom level for the mesh}
}
\description{
Sets up a simple htmlwidget displaying a mesh with
brainbrowser
}
