% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{fig_to_html}
\alias{fig_to_html}
\title{Figure to html}
\usage{
fig_to_html(fig, unit_w = "px", unit_h = "px")
}
\arguments{
\item{fig}{the layout}

\item{unit_w}{the css unit to use for the widths}

\item{unit_h}{the css unit to use for the heights}
}
\description{
Convert a layout to an html page full of widgets.
}
