% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\docType{data}
\name{height_l}
\alias{height_l}
\title{A lens into the height of a layout object}
\format{An object of class \code{lens} of length 2.}
\usage{
height_l
}
\description{
A lens into the height of a layout object
}
\keyword{datasets}
