% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbrainbrowser-package.R
\docType{package}
\name{rbrainbrowser-package}
\alias{rbrainbrowser}
\alias{rbrainbrowser-package}
\title{rbrainbrowser: Brainbrowser html widgets}
\description{
Render surface meshes as htmlwidgets with brainbrowser
}
\keyword{internal}
