% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{reify_fig}
\alias{reify_fig}
\title{Materialize a figure}
\usage{
reify_fig(fig, height = 400, width = 600)
}
\description{
Materialize the figure by recursively spreading the layout to
fit a fixed width and height
}
