% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{snap_page}
\alias{snap_page}
\title{Take a screenshot}
\usage{
snap_page(site, file = "test.png", port = 4869L, height = 800,
  width = 1280, version = NULL)
}
\arguments{
\item{site}{the url to image}

\item{file}{a png file to save the output}

\item{port}{an integer port number for the chrome driver}

\item{height}{a height in pixels for the page viewport}

\item{width}{a width in pixels for the page viewport}

\item{version}{the version of the chrome driver to use}
}
\description{
Take a snapshot of an html page using the chrome drive
}
