% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{spread_heights}
\alias{spread_heights}
\title{Spread a height across a layout}
\usage{
spread_heights(fig, height)
}
\arguments{
\item{fig}{A layout}

\item{height}{The height to fill}
}
\description{
Expand layout objects proportionality to fit a fixed height.
}
