% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{spread_widths}
\alias{spread_widths}
\title{Spread a width across a layout}
\usage{
spread_widths(fig, width)
}
\arguments{
\item{fig}{A layout}

\item{width}{The width to fill}
}
\description{
Expand layout objects proportionality to fit a fixed width.
}
