% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\name{unify_data}
\alias{unify_data}
\title{Unify data}
\usage{
unify_data(fig, data = NULL)
}
\arguments{
\item{fig}{a layout}

\item{data}{the reference data}
}
\description{
Fill in missing layout data from a reference list of arguments
to `bb`
}
