% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/brainbrowser.R
\docType{data}
\name{width_l}
\alias{width_l}
\title{A lens into the width of a layout object}
\format{An object of class \code{lens} of length 2.}
\usage{
width_l
}
\description{
A lens into the width of a layout object
}
\keyword{datasets}
